refactor: connection and logic
