+++ durable-objects/wrangler.jsonc
@@ -5,7 +5,7 @@
 {
 	"$schema": "node_modules/wrangler/config-schema.json",
 	"name": "durable-objects",
-	"main": "src/ws1.ts",
+	"main": "src/index.ts",
 	"compatibility_date": "2025-08-13",
 	"migrations": [
 		{
